Guess The Sum of the Die Game
Console based game.
This is a three player guessing game. The goal of the game is to guess the sum of the two die are being rolled.  Each player start off with 2000 pesos. Each player would bet an amount for every guess. The game ends when only one player remains.